Let Timestamps plugin use Tumblr API v2
I tried to rewrite the Timestamps plugin (or at least the part that was
throwing insecure content errors) to use Tumblr API v2.

Collateral damage: adding permissions for all HTTPS sites, having to use
a Tumblr API key (that MUST be changed; this one is the example one
stolen from the documentation).