Let Blog Tracker plugin use Tumblr API v2
Again, this is for purposes of switching to HTTPS, and the API key must
be changed.
This time I noticed that the version currently in production differs
quite a bit from this version; however, the changes that must be made
are relatively simple.
Also, changing the avatar URL to HTTPS is currently pointless because it
simply 301's to an image on media.tumblr.com, served over HTTP. Oh well.